requirements.txt
# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0

# Database
psycopg2-binary==2.9.9

# Configuration
python-dotenv==1.0.0

# Serialization/Validation
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Security
bcrypt==4.1.2

# Production Server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Development Tools (optional)
# Uncomment for development
# black==23.12.1
# flake8==7.0.0
# pylint==3.0.3